# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Carpoolink API
  version: 0.1.0
  description: 사용자 식별이 필요한 API는 `x-user-id` 헤더가 필요합니다.
servers:
  - url: http://localhost:3001
...
paths:
...
  /surveys:
    get:
      summary: 사전 설문의 질문과 선택지 조회
      operationId: getSurveyQuestions
      tags: [Survey]
      responses:
        '200':
          description: 설문 질문과 선택지 목록 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveyQuestionListResponse'
              example:
                surveyQuestions:
                  - surveyQuestionId: '1'
                    code: GOAL
                    question: 멘토링 목표는 무엇인가요?
                    displayOrder: 1
                    options:
                      - surveyQuestionOptionId: '11'
                        code: A
                        label: 취업 준비
                        displayOrder: 1
                      - surveyQuestionOptionId: '12'
                        code: B
                        label: 실무 역량 강화
                        displayOrder: 2
        '500':
          $ref: '#/components/responses/InternalServerError'
  /surveys/submit:
    post:
      summary: 사전 설문 결과 제출 및 유형 판정
      operationId: submitSurvey
      tags: [Survey]
      parameters:
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SurveySubmitRequest'
            example:
              answers:
                goal: A
                style: B
                preference: A
                focus: B
      responses:
        '200':
          description: 사전 설문 결과 및 유형 저장 및 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveySubmitResponse'
              example:
                result:
                  surveyResultId: '3'
                  combinationCode: ABAB
                  title: 공감형 전략자
                answers:
                  - question: GOAL
                    answer: 취업 준비
                  - question: TIME
                    answer: 주말 중심
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: 멘티가 아닌 사용자가 설문 제출 시도
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: 멘티만 설문에 참여할 수 있습니다.
        '404':
          description: 매칭되는 설문 유형 미존재 또는 사용자 미존재
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    XUserIdHeader:
      name: x-user-id
      in: header
      required: true
      description: 인증/식별에 사용되는 사용자 ID(BigInt 문자열)
      schema:
        type: string
        pattern: '^[0-9]+$'
      example: '1001'
...
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: x-user-id header가 필요합니다.
...
    InternalServerError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Internal Server Error

# ------------------------------------------------------------------------------------------
# ------------------------------------------------------------------------------------------
# ------------------------------------------------------------------------------------------
# 요청/응답에 사용되는 데이터 모델 정의
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    SurveyOption:
      type: object
      properties:
        surveyQuestionOptionId:
          type: string
          nullable: true
        code:
          type: string
        label:
          type: string
        displayOrder:
          type: integer
          nullable: true
      required: [code, label]
    SurveyQuestion:
      type: object
      properties:
        surveyQuestionId:
          type: string
          nullable: true
        code:
          type: string
        question:
          type: string
          nullable: true
        displayOrder:
          type: integer
          nullable: true
        options:
          type: array
          items:
            $ref: '#/components/schemas/SurveyOption'
      required: [code, options]
    SurveyQuestionListResponse:
      type: object
      properties:
        surveyQuestions:
          type: array
          items:
            $ref: '#/components/schemas/SurveyQuestion'
      required: [surveyQuestions]
    SurveySubmitRequest:
      type: object
      properties:
        answers:
          type: object
          description: 질문 코드 -> 선택지 코드 매핑
          additionalProperties:
            type: string
      required: [answers]
    SurveyResult:
      type: object
      description: DB의 surveyResult 전체 객체를 반환하며 필드는 확장될 수 있습니다.
      properties:
        surveyResultId:
          type: string
          nullable: true
        combinationCode:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
      additionalProperties: true
    SurveyAnswerSummary:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
      required: [question, answer]
    SurveySubmitResponse:
      type: object
      properties:
        result:
          $ref: '#/components/schemas/SurveyResult'
        answers:
          type: array
          items:
            $ref: '#/components/schemas/SurveyAnswerSummary'
      required: [result, answers]
...